build awesome books with es6 modules
