+++ index.html
@@ -58,8 +58,6 @@
       integrity="sha384-ENjdO4Dr2bkBIFxQpeoTz1HIcje39Wm4jDKdf19U8gI4ddQ3GYNS7NTKfAdVQSZe"
       crossorigin="anonymous"
     ></script>
-    <script src="js/date.js"></script>
-    <!-- <script src="js/script-class.js"></script> -->
-    <script src="js/spa.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
