--- js/date.js
@@ -1,1 +1,0 @@
-document.querySelector('time').innerText = new Date();
