+++ modules/DOMLoader.js
@@ -1,0 +1,4 @@
+const mainContainer = document.querySelector('main');
+const navLinks = document.querySelectorAll('.nav-link');
+
+export { mainContainer, navLinks };
